graph cleanup
