+++ src/editor/editor.cpp
@@ -11,6 +11,8 @@
 #include "input_state.hpp"
 #include "time.hpp"
 
+#include "graph/graph_window.hpp"
+#include "graph/node_properties.hpp"
 #include "graphics/icon_set.hpp"
 #include "operations/operation_stack.hpp"
 #include "renderers/id_renderer.hpp"
@@ -36,7 +38,6 @@
 #include "windows/commands_window.hpp"
 #include "windows/composer_window.hpp"
 #include "windows/debug_view_window.hpp"
-#include "windows/graph_window.hpp"
 #include "windows/icon_browser.hpp"
 #include "windows/layers_window.hpp"
 #include "windows/network_window.hpp"
@@ -593,6 +594,7 @@ public:
                 m_clipboard_window       = std::make_unique<Clipboard_window                >(*m_imgui_renderer.get(), *m_imgui_windows.get(),  m_editor_context);
                 m_commands_window        = std::make_unique<Commands_window                 >(*m_imgui_renderer.get(), *m_imgui_windows.get(),  m_editor_context);
                 m_graph_window           = std::make_unique<Graph_window                    >(*m_commands.get(),       *m_imgui_renderer.get(), *m_imgui_windows.get(),  m_editor_context, *m_editor_message_bus.get());
+                m_node_properties_window = std::make_unique<Node_properties_window          >(*m_imgui_renderer.get(), *m_imgui_windows.get(),  m_editor_context);
                 m_icon_browser           = std::make_unique<Icon_browser                    >(*m_imgui_renderer.get(), *m_imgui_windows.get(),  m_editor_context);
                 m_sheet_window           = std::make_unique<Sheet_window                    >(*m_commands.get(),       *m_imgui_renderer.get(), *m_imgui_windows.get(),  m_editor_context, *m_editor_message_bus.get());
                 m_layers_window          = std::make_unique<Layers_window                   >(*m_imgui_renderer.get(), *m_imgui_windows.get(),  m_editor_context);
@@ -1132,6 +1134,7 @@ public:
     std::unique_ptr<Clipboard_window                >        m_clipboard_window;
     std::unique_ptr<Commands_window                 >        m_commands_window;
     std::unique_ptr<Graph_window                    >        m_graph_window;
+    std::unique_ptr<Node_properties_window          >        m_node_properties_window;
     std::unique_ptr<Icon_browser                    >        m_icon_browser;
     std::unique_ptr<Sheet_window                    >        m_sheet_window;
     std::unique_ptr<Layers_window                   >        m_layers_window;
