+++ src/editor/graph/constant.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+#include "graph/shader_graph_node.hpp"
+
+namespace editor {
+
+class Constant : public Shader_graph_node
+{
+public:
+    Constant();
+
+    void evaluate(Shader_graph&) override;
+    void imgui   () override;
+
+    Payload m_payload;
+};
+
+} // namespace editor
